requirements.txt
annotated-types==0.6.0
greenlet==3.0.3
pydantic==2.6.1
pydantic_core==2.16.2
python-dotenv==1.0.1
SQLAlchemy==2.0.27
typing_extensions==4.9.0